LayoutTests/platform/blackberry/editing/pasteboard/paste-noscript-expected.txt

CONSOLE MESSAGE: line 38: TypeError: Result of expression 'document.getElementById("pastehere").childNodes(3)' [null] is not an object.
This test copies all the elements containing event handlers and javascript urls, pastes them in an editable area and verifies that no script, handlers or javascript urls are copied.
Hello 
CNN Hello 
This is a form

<button id="button1" onclick="sayHello()" ondblclick="sayHello()" style="width: 100px;">Hello</button>
undefined
<a id="anchor1" href="http://www.cnn.com/">CNN</a>
